refactor: revised navbar container widths (#36)
The navbar did not resize to take in the container widths that resize according the the responsive behaviors set in other sections of the webpage, this change implements media breakpoints that are used elsewhere so that the behavior is consistent.

[squash] Storybook build failure fix
Removed file references for elements within storybook build that would fail during build.
Amended the tests so that they follow the alternate code design