bug: removed mui due to incompatibility
The MUI libraries cause an issue when attempting to use the TailwindCSS styling library in production. With attempts to fix the layering order, all unsuccessful, it is best to remove MUI as Tailwind CSS has some resources and components that could be used to replace the functionality provided by MUI component library.